+++ src/test/java/org/duckstudy/model/lotto/LottoTest.java
@@ -1,8 +1,9 @@
-package org.duckstudy.model;
+package org.duckstudy.model.lotto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import org.duckstudy.model.lotto.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
